Se corrige interfaces del patron observer, se agrega instancia para Aparcadero como suscriptor, y se arregla el modelo de la base de datos

+++ simulation-bus/migrations/20231017170832_db_busrta.js
@@ -41,7 +41,7 @@ const up = function (knex) {
       table.integer("ruta").unsigned();
       table.foreign("ruta").references("ruta.id");
 
-      table.float("indice_ruta").defaultTo(0);
+      table.integer("indice_ruta").defaultTo(0);
     })
     .createTable("simulacion", function (table) {
       table.increments("id").primary();
